+++ src/lib/pytornado/objects/settings.py
@@ -49,12 +49,11 @@ _DEFAULT_PLOT_DICT = {
     'results': (_PLOT_OPTIONS, dict),
 }
 
-_SAVE_RESULTS = [
-    "NO_global",
-    "NO_panelwise",
-    "NO_loads_with_undeformed_mesh",
-    "NO_loads_with_deformed_mesh"
-]
+_DEFAULT_SAVE_DICT = {
+    'global': (False, bool),
+    'panelwise': (False, bool),
+    'aeroperformance': (False, bool),
+}
 
 DEFAULT_SETTINGS = {
     'aircraft': (None, str),
@@ -62,7 +61,7 @@ DEFAULT_SETTINGS = {
     'deformation': (None, (None, str)),
     'vlm_autopanels_c': (4, int),
     'vlm_autopanels_s': (4, int),
-    'save_results': (_SAVE_RESULTS, list),
+    'save_results': (_DEFAULT_SAVE_DICT, dict),
     'plot': (_DEFAULT_PLOT_DICT, dict),
     # Underscore settings are "hidden" settings that generally shouldn't be changed
     '_do_normal_rotations': (True, bool),
@@ -181,9 +180,9 @@ class Settings:
         self.paths.add_subpath(uid_parent='d_settings', uid='f_settings', path=f"{self.project_basename}.json")
         self.paths.add_subpath(uid_parent='d_state', uid='f_state', path=f"{self.settings['state']}")
         # Output files
-        self.paths.add_subpath(uid_parent='d_results', uid='f_results_global', path=f"{self.project_basename}_global.json")
-        self.paths.add_subpath(uid_parent='d_results', uid='f_results_panelwise', path=f"{self.project_basename}_global.json")
-        self.paths.add_subpath(uid_parent='d_results', uid='f_results_apm_global', path=f"{self.project_basename}_APM.json")
+        self.paths.add_subpath(uid_parent='d_results', uid='f_results_global', path=f"global.json")  # File namespaced with folder
+        self.paths.add_subpath(uid_parent='d_results', uid='f_results_panelwise', path=f"panelwise.dat")  # File namespaced with folder
+        self.paths.add_subpath(uid_parent='d_results_TOP', uid='f_results_apm_global', path=f"{self.project_basename}_aeroperformance.json")
 
     def _check_aircraft_file_type(self):
         """Check whether aircraft is a CPACS or a JSON file"""